pyproject.toml
[tool.poetry]
name = "python-tcp-server"
version = "0.1.0"
description = ""
authors = ["GabsMarcelino <gmarcelino@tractian.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.13.0"
requests = "^2.32.5"
rich = "^14.2.0"
pillow = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
